adding a folder for each section on cart theme 1
