+++ EcoMarket/Scenes/TabBar/Cart/CartTheme1/CartProductSection/CartProductCollectionViewCell.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class PrroductCollectionViewCell: UICollectionViewCell {
+class CartProductCollectionViewCell: UICollectionViewCell {
     // MARK: - Outlets
     //
     @IBOutlet weak var containerView: UIStackView!
