+++ EcoMarket/Scenes/TabBar/Cart/CartTheme1/CartProductSection/CartProductsSection.swift
@@ -1,0 +1,84 @@
+//
+//  CartProductsSection.swift
+//  EcoMarket
+//
+//  Created by Ibrahim Nasser Ibrahim on 31/01/2024.
+//
+
+import UIKit
+
+class CartProductsSection: CollectionSection {
+    typealias ItemsType = CartModel
+    var items: [CartModel] = []
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell: CartProductCollectionViewCell = collectionView.dequeue(indexPath: indexPath) else {
+            Logger.log("Can't dequeue ProductsCollectionViewCell", category: \.default, level: .fault)
+            return UICollectionViewCell()
+        }
+        cell.setup(product: CartModel.products[indexPath.row])
+        return cell
+    }
+    
+    func numberOfItems() -> Int {
+        items.count
+    }
+    
+    let interItemSpacing: CGFloat = 5
+    let padding: CGFloat = 15.0
+    let height: CGFloat = 130
+    
+    func sectionLayout() -> NSCollectionLayoutSection {
+        // Item
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
+                                              heightDimension: .fractionalHeight(1.0))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        // Group
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .absolute(height))
+        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
+        group.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: padding, bottom: 0, trailing: padding)
+        
+        // Header
+        let header = createHeader()
+        
+        // Section
+        let section = NSCollectionLayoutSection(group: group)
+        section.boundarySupplementaryItems = [header]
+        section.contentInsets.bottom = 40
+        return section
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        guard let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind,
+                                                                           withReuseIdentifier: Header.identifier,
+                                                                           for: indexPath) as? Header else {
+            Logger.log("Failed to get header view", category: \.default, level: .fault)
+            return UICollectionReusableView()
+        }
+        header.setTitle("My Cart")
+        return header
+    }
+    
+    func registerCell(in collectionView: UICollectionView) {
+        collectionView.registerNib(CartProductCollectionViewCell.self)
+    }
+    private func createHeader() -> NSCollectionLayoutBoundarySupplementaryItem {
+        let headerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
+                                                heightDimension: .estimated(50))
+        let header = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: headerSize,
+            elementKind: Header.elementKind,
+            alignment: .top
+        )
+        
+        header.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 20, bottom: 0, trailing: 0)
+        return header
+    }
+    func registerSupplementaryView(in collectionView: UICollectionView) {
+        collectionView.register(Header.self,
+                                forSupplementaryViewOfKind: Header.elementKind,
+                                withReuseIdentifier: Header.identifier)
+    }
+}
